optimize: extract shared scripts/lib/okf.mjs (agent 1, dedup-dry)
okf-index / okf-validate / okf-backfill each redefined RESERVED, walk(),
parseFrontmatter(), the exclude predicate, and the config path. Hoist them into
one scripts/lib/okf.mjs (walk takes an optional isExcluded; parseFrontmatter
standardizes on {fields}|null, with okf-index adapting its one call site).
Verified byte-identical output: validate 95 concepts OK, index --check OK,
backfill --dry nothing-to-do.
